Promote location ports to a new core:location module
GeofenceRegistrar and LocationProvider are not reminder-specific.
LocationSnoozeProcessor was the only in-module consumer; SnoozeSettingsFragment
(feature:ui) and Autostart (app) imported them too. Lift the two port
interfaces plus the GMS/NoOp flavor-specific adapters into a new
core:location module that owns the full/foss flavor split.

Cross-module re-entry for the geofence transition stays on the OS broadcast
bus to avoid a core:location -> feature:reminders compile dependency. The
GeofenceBroadcastReceiver in core:location forwards via
ProcessorCode.GeofenceEntered to ReminderProcessorBroadcastReceiver in
feature:reminders, which calls the new ReminderCommandBus.restoreLocationSnoozes
method backed by LocationSnoozeProcessor.

Update gradle dependencies in :app, :feature:ui, and :feature:reminders to
include :core:location; restore the distribution flavour declaration on
:feature:reminders so variant resolution matches the flavoured :core:location.
GeofenceBroadcastReceiverTest is rewritten to verify the broadcast forwarding
behaviour instead of the old direct-injection one.
