+++ app/build.gradle.kts
@@ -125,6 +125,7 @@ dependencies {
     implementation(project(":core:common"))
     implementation(project(":core:datastore"))
     implementation(project(":core:ui"))
+    implementation(project(":core:location"))
     implementation(project(":feature:reminders"))
     implementation(project(":feature:ui"))
     // Compose
