+++ app/src/test/java/com/futsch1/medtimer/location/PreferencesDataSourceHomeLocationTest.kt
@@ -1,4 +1,4 @@
-package com.futsch1.medtimer.feature.reminders.location
+package com.futsch1.medtimer.location
 
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
