+++ core/location/src/foss/java/com/futsch1/medtimer/core/location/NoOpLocationProvider.kt
@@ -1,0 +1,11 @@
+package com.futsch1.medtimer.core.location
+
+import android.location.Location
+import javax.inject.Inject
+
+class NoOpLocationProvider @Inject constructor() : LocationProvider {
+    override fun getCurrentLocation(onSuccess: (Location?) -> Unit, onFailure: () -> Unit): () -> Unit {
+        onFailure()
+        return {}
+    }
+}
