+++ core/location/src/main/java/com/futsch1/medtimer/core/location/LocationProvider.kt
@@ -1,0 +1,7 @@
+package com.futsch1.medtimer.core.location
+
+import android.location.Location
+
+fun interface LocationProvider {
+    fun getCurrentLocation(onSuccess: (Location?) -> Unit, onFailure: () -> Unit): () -> Unit
+}
