--- feature/reminders/src/foss/java/com/futsch1/medtimer/feature/reminders/location/NoOpGeofenceRegistrar.kt
@@ -1,13 +1,0 @@
-package com.futsch1.medtimer.feature.reminders.location
-
-import javax.inject.Inject
-
-class NoOpGeofenceRegistrar @Inject constructor() : GeofenceRegistrar {
-    override val isSupported: Boolean = false
-    override fun isLocationServiceAvailable() = false
-    override fun registerHomeGeofence(onSuccess: (() -> Unit)?, onFailure: (() -> Unit)?) = false
-    override fun unregisterHomeGeofence() {
-        // Intentionally empty
-    }
-    override fun hasRequiredPermissions() = false
-}
