--- feature/reminders/src/foss/java/com/futsch1/medtimer/feature/reminders/location/NoOpLocationProvider.kt
@@ -1,11 +1,0 @@
-package com.futsch1.medtimer.feature.reminders.location
-
-import android.location.Location
-import javax.inject.Inject
-
-class NoOpLocationProvider @Inject constructor() : LocationProvider {
-    override fun getCurrentLocation(onSuccess: (Location?) -> Unit, onFailure: () -> Unit): () -> Unit {
-        onFailure()
-        return {}
-    }
-}
