--- feature/reminders/src/full/AndroidManifest.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application>
-        <receiver
-            android:name=".location.GeofenceBroadcastReceiver"
-            android:enabled="true"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="com.futsch1.medtimer.GEOFENCE_TRANSITION" />
-            </intent-filter>
-        </receiver>
-    </application>
-
-</manifest>
