--- feature/reminders/src/full/java/com/futsch1/medtimer/feature/reminders/location/GmsGeofenceRegistrar.kt
@@ -1,117 +1,0 @@
-package com.futsch1.medtimer.feature.reminders.location
-
-import android.Manifest
-import android.app.PendingIntent
-import android.content.Context
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.os.Build
-import android.util.Log
-import androidx.core.content.ContextCompat
-import com.futsch1.medtimer.core.common.LogTags
-import com.futsch1.medtimer.core.datastore.PreferencesDataSource
-import com.futsch1.medtimer.core.domain.model.HomeLocation
-import com.google.android.gms.common.ConnectionResult
-import com.google.android.gms.common.GoogleApiAvailability
-import com.google.android.gms.location.Geofence
-import com.google.android.gms.location.GeofencingClient
-import com.google.android.gms.location.GeofencingRequest
-import dagger.hilt.android.qualifiers.ApplicationContext
-import javax.inject.Inject
-
-class GmsGeofenceRegistrar @Inject constructor(
-    @param:ApplicationContext private val context: Context,
-    private val geofencingClient: GeofencingClient,
-    private val preferencesDataSource: PreferencesDataSource,
-    private val googleApiAvailability: GoogleApiAvailability
-) : GeofenceRegistrar {
-
-    override val isSupported: Boolean = true
-
-    override fun isLocationServiceAvailable(): Boolean =
-        googleApiAvailability.isGooglePlayServicesAvailable(context) == ConnectionResult.SUCCESS
-
-    override fun registerHomeGeofence(onSuccess: (() -> Unit)?, onFailure: (() -> Unit)?): Boolean {
-        if (!isLocationServiceAvailable()) {
-            Log.w(LogTags.LOCATION, "Google Play Services unavailable, cannot register home geofence")
-            return false
-        }
-        val homeLocation = preferencesDataSource.preferences.value.homeLocation ?: run {
-            Log.w(LogTags.LOCATION, "No home location saved, cannot register geofence")
-            return false
-        }
-        if (!hasRequiredPermissions()) {
-            Log.w(LogTags.LOCATION, "Location permissions not granted, cannot register geofence")
-            return false
-        }
-
-        if (!isValid(homeLocation)) {
-            Log.w(LogTags.LOCATION, "Invalid home location parameters: $homeLocation")
-            return false
-        }
-
-        val geofence = Geofence.Builder()
-            .setRequestId(GeofenceRegistrar.GEOFENCE_ID)
-            .setCircularRegion(homeLocation.latitude, homeLocation.longitude, homeLocation.radiusMeters)
-            .setExpirationDuration(Geofence.NEVER_EXPIRE)
-            // Quick loitering delay, as the geofence will be unregistered as soon as the snoozed events are fired.
-            // So if we are already inside the fence, this should trigger soon.
-            .setLoiteringDelay(30_000)
-            .setTransitionTypes(Geofence.GEOFENCE_TRANSITION_ENTER or Geofence.GEOFENCE_TRANSITION_DWELL)
-            .build()
-
-        val request = GeofencingRequest.Builder()
-            .addGeofence(geofence)
-            .build()
-
-        return try {
-            geofencingClient.addGeofences(request, buildGeofencePendingIntent())
-                .addOnSuccessListener {
-                    Log.i(LogTags.LOCATION, "Home geofence registered successfully")
-                    onSuccess?.invoke()
-                }
-                .addOnFailureListener {
-                    Log.e(LogTags.LOCATION, "Failed to add home geofence: ${it.message}")
-                    onFailure?.invoke()
-                }
-            true
-        } catch (e: SecurityException) {
-            Log.e(LogTags.LOCATION, "Security exception registering home geofence: ${e.message}")
-            false
-        }
-    }
-
-    override fun unregisterHomeGeofence() {
-        geofencingClient.removeGeofences(listOf(GeofenceRegistrar.GEOFENCE_ID))
-            .addOnSuccessListener { Log.i(LogTags.LOCATION, "Home geofence removed") }
-            .addOnFailureListener { Log.w(LogTags.LOCATION, "Failed to remove home geofence: ${it.message}") }
-    }
-
-    private fun buildGeofencePendingIntent(): PendingIntent {
-        val intent = Intent(context, GeofenceBroadcastReceiver::class.java)
-        return PendingIntent.getBroadcast(
-            context, 0, intent,
-            PendingIntent.FLAG_MUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
-        )
-    }
-
-    private fun isValid(homeLocation: HomeLocation): Boolean {
-        return homeLocation.latitude in -90.0..90.0 &&
-                homeLocation.longitude in -180.0..180.0 &&
-                homeLocation.radiusMeters > 0
-    }
-
-    override fun hasRequiredPermissions(): Boolean {
-        val fineLocation = ContextCompat.checkSelfPermission(
-            context, Manifest.permission.ACCESS_FINE_LOCATION
-        ) == PackageManager.PERMISSION_GRANTED
-        val backgroundLocation = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-            ContextCompat.checkSelfPermission(
-                context, Manifest.permission.ACCESS_BACKGROUND_LOCATION
-            ) == PackageManager.PERMISSION_GRANTED
-        } else {
-            true
-        }
-        return fineLocation && backgroundLocation
-    }
-}
