+++ feature/reminders/src/main/java/com/futsch1/medtimer/feature/reminders/SnoozeProcessor.kt
@@ -3,7 +3,7 @@ package com.futsch1.medtimer.feature.reminders
 import android.util.Log
 import com.futsch1.medtimer.core.common.LogTags
 import com.futsch1.medtimer.core.datastore.PersistentDataDataSource
-import com.futsch1.medtimer.feature.reminders.location.GeofenceRegistrar
+import com.futsch1.medtimer.core.location.GeofenceRegistrar
 import com.futsch1.medtimer.feature.reminders.notificationData.ReminderNotificationData
 import com.futsch1.medtimer.feature.reminders.notificationData.toPendingSnooze
 import java.time.Instant
