--- feature/reminders/src/main/java/com/futsch1/medtimer/feature/reminders/location/LocationProvider.kt
@@ -1,7 +1,0 @@
-package com.futsch1.medtimer.feature.reminders.location
-
-import android.location.Location
-
-fun interface LocationProvider {
-    fun getCurrentLocation(onSuccess: (Location?) -> Unit, onFailure: () -> Unit): () -> Unit
-}
